refac: tiered fees component
