+++ src/components/nft-dashboard/Balance/components/SendForm/SendForm.styles.ts
@@ -13,20 +13,6 @@ export const TextRow = styled.div`
   gap: 1rem;
 `;
 
-export const SubCard = styled(BaseCard)<{ $isMobile?: boolean }>`
-  width: 30%;
-  ${(props) =>
-    props.$isMobile &&
-    css`
-      width: 100%;
-    `}
-  background-color: var(--additional-background-color);
-  cursor: pointer;
-  box-shadow: 0px 0px 10px 0px var(--shadow-color);
-  .ant-card-body {
-    padding: 1rem 2rem;
-  }
-`;
 export const SendBody = styled(BaseRow)`
   padding-bottom: 1rem;
 `;
@@ -45,28 +31,16 @@ export const FormHeader = styled.span`
   padding-bottom: 1rem;
 `;
 
-export const SubCardHeader = styled.span`
-  font-size: 1.5rem;
-`;
-
 export const InputHeader = styled.span`
   font-size: 1.5rem;
 `;
 
-export const SubCardAmount = styled.span`
-  font-size: 1.5rem;
-`;
-export const SubCardContent = styled.div`
-  font-size: 1.3rem;
-  height: 100%;
+export const RBFWrapper = styled.div`
   display: flex;
-  justify-content: space-around;
-  flex-direction: column;
+  flex-direction: row;
+  gap: 1rem;
   align-items: center;
-  text-align: center;
-  gap: 3rem;
-  padding-top: 1rem;
-  padding-bottom: 1rem;
+  justify-content: center;
 `;
 
 export const InputWrapper = styled.div`
@@ -74,22 +48,6 @@ export const InputWrapper = styled.div`
   min-width: 25vw;
   flex-direction: column;
   gap: 0.5rem;
-`;
-export const TiersRow = styled.div`
-  display: flex;
-  flex-direction: row;
-  gap: 1rem;
-  justify-content: space-around;
-`;
-
-export const TiersCol = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  justify-content: space-around;
-`;
-export const SendFormButton = styled(BaseButton)`
-  width: 100%;
 `;
 export const TiersContainer = styled.div`
   display: flex;
@@ -107,6 +65,10 @@ export const TiersContainer = styled.div`
     border: 1px solid var(--error-color);
   }
 `;
+export const SendFormButton = styled(BaseButton)`
+  width: 100%;
+`;
+
 export const BalanceInfo = styled.small`
   color: var(--subtext-color);
 `;
@@ -125,19 +87,16 @@ export const AddressText = styled.span`
   text-decoration: underline;
   color: var(--text-main-color);
 `;
-
-export const RateValueWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-`;
-export const RateValue = styled.span`
-  color: green;
-`;
-export const RBFWrapper = styled.div`
+export const TiersRow = styled.div`
   display: flex;
   flex-direction: row;
   gap: 1rem;
-  align-items: center;
-  justify-content: center;
+  justify-content: space-around;
+`;
+
+export const TiersCol = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  justify-content: space-around;
 `;
